move playbar controls to useSpeechActions hook
